Skip files when available

+++ filedownloader.py
@@ -3,6 +3,9 @@ import requests
 
 
 def download_file(url, filename):
+    if os.path.isfile(filename):
+        print("File exists. Skipping...")
+        return
     r = requests.get(url, stream=True)
     with open(filename,'wb') as f:
         for chunk in r.iter_content(chunk_size=1024):
